Revert "added limit switches to arm"
This reverts commit 9206567e9de239fe5b7cd95cff51993ef5c54279, reversing
changes made to c1628969d235108da7ed8e06ea2e11203f125951.

+++ src/main/java/frc/robot/subsystems/ClawSubsystem.java
@@ -28,7 +28,7 @@ public class ClawSubsystem extends SubsystemBase {
     // This method will be called once per scheduler run
   }
 
-  public void rotate(double direction)
+  public void rotate(int direction)
   {
     rotationMotor.set(direction);
   }
